Create initial Repo (Life Canvas)
