--- .github/workflows/template_change_test.yml
@@ -1,47 +1,0 @@
-name: Rename Template
-
-on: pull_request
-
-jobs:
-  rename-template:
-    strategy:
-      matrix:
-        useHiltDependencies: [true, false]
-        useRoomDependencies: [true, false]
-        useRetrofitDependencies: [true, false]
-        usePaparazziDependencies: [true, false]
-
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Set useHiltDependencies
-        run: |
-          sed -i '' 's/useHiltDependencies.* : true/useHiltDependencies : ${{ matrix.useHiltDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set useRoomDependencies
-        run: |
-          sed -i '' 's/useRoomDependencies.* : true/useRoomDependencies : ${{ matrix.useRoomDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set useRetrofitDependencies
-        run: |
-          sed -i '' 's/useRetrofitDependencies.* : true/useRetrofitDependencies : ${{ matrix.useRetrofitDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set usePaparazziDependencies
-        run: |
-          sed -i '' 's/usePaparazziDependencies.* : true/usePaparazziDependencies : ${{ matrix.usePaparazziDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set Up JDK
-        uses: actions/setup-java@v4
-        with:
-          distribution: 'zulu'
-          java-version: 17
-
-      - name: Setup Gradle
-        uses: gradle/gradle-build-action@v2
-
-      - name: Rename
-        run: ./gradlew renameTemplate
-
-      - name: Build
-        run: ./gradlew build
