+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.CustomSplashScreenTheme">
 
         <activity
-            android:name="template.MainActivity"
+            android:name="info.hridoydas.lifecanvas.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.AppTheme.NoActionBar"
             android:windowSoftInputMode="adjustResize">
