+++ app/src/main/java/info/hridoydas/lifecanvas/MainActivity.kt
@@ -1,0 +1,157 @@
+/*
+* MIT License
+*
+* Copyright (c) 2024 Hridoy Chandra Das
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in all
+* copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*
+*/
+package info.hridoydas.lifecanvas
+
+import android.animation.ObjectAnimator
+import android.os.Bundle
+import android.view.View
+import android.view.animation.OvershootInterpolator
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.viewModels
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.ui.graphics.Color
+import androidx.core.animation.doOnEnd
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.core.view.WindowCompat
+import androidx.navigation.compose.rememberNavController
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+import dagger.hilt.android.AndroidEntryPoint
+import info.hridoydas.lifecanvas.common.DURATION
+import info.hridoydas.lifecanvas.common.VALUES_X
+import info.hridoydas.lifecanvas.common.VALUES_Y
+import info.hridoydas.lifecanvas.common.utils.RootUtil
+import info.hridoydas.lifecanvas.theme.LifeCanvasTheme
+import info.hridoydas.lifecanvas.theme.splashScreen.SplashViewModel
+import info.hridoydas.lifecanvas.ui.MainAnimationNavHost
+import timber.log.Timber
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    companion object {
+        private val Tag = MainActivity::class.java.simpleName
+    }
+
+    private val splashViewModel: SplashViewModel by viewModels()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        configureEdgeToEdgeWindow()
+
+        // Check Rooted Device
+        if (RootUtil.isDeviceRooted()) {
+            Timber.tag(Tag).e("onCreate - Rooted device.")
+            finish()
+            return
+        }
+
+        Timber.tag(Tag).d("onCreate")
+
+        installSplashScreen().apply {
+            setKeepOnScreenCondition {
+                !splashViewModel.isLoading.value
+            }
+            setOnExitAnimationListener { screen ->
+                val zoomX = ObjectAnimator.ofFloat(
+                    screen.iconView,
+                    View.SCALE_X,
+                    VALUES_X,
+                    VALUES_Y,
+                )
+                zoomX.interpolator = OvershootInterpolator()
+                zoomX.duration = DURATION
+                zoomX.doOnEnd { screen.remove() }
+
+                val zoomY = ObjectAnimator.ofFloat(
+                    screen.iconView,
+                    View.SCALE_Y,
+                    VALUES_X,
+                    VALUES_Y,
+                )
+                zoomY.interpolator = OvershootInterpolator()
+                zoomY.duration = DURATION
+                zoomY.doOnEnd { screen.remove() }
+
+                zoomX.start()
+                zoomY.start()
+            }
+        }
+        // splashViewModel.checkStartScreen() { route -> }
+
+        setContent {
+            ConfigureTransparentSystemBars()
+
+            LifeCanvasTheme {
+                Surface(
+                    color = MaterialTheme.colorScheme.background,
+                ) {
+                    val navController = rememberNavController()
+                    MainAnimationNavHost(navController)
+                }
+            }
+        }
+    }
+
+    /**
+     * Configures our [MainActivity] window so that it reaches edge to edge of the device, meaning
+     * content can be rendered underneath the status and navigation bars.
+     *
+     * This method works hand in hand with [ConfigureTransparentSystemBars], to make sure content
+     * behind these bars is visible.
+     *
+     * Keep in mind that if you need to make sure your content padding doesn't clash with the status bar text/icons,
+     * you can leverage modifiers like `windowInsetsPadding()` and `systemBarsPadding()`. For more information,
+     * read the Compose WindowInsets docs: https://developer.android.com/reference/kotlin/androidx/compose/foundation/layout/WindowInsets
+     */
+    private fun configureEdgeToEdgeWindow() {
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+    }
+
+    @Composable
+    private fun ConfigureTransparentSystemBars() {
+        val systemUiController = rememberSystemUiController()
+        val useDarkIcons = !isSystemInDarkTheme()
+
+        DisposableEffect(systemUiController, useDarkIcons) {
+            systemUiController.setSystemBarsColor(
+                color = Color.Transparent,
+                darkIcons = useDarkIcons,
+            )
+
+            onDispose { }
+        }
+    }
+}
+
+@Composable
+fun Greeting(name: String) {
+    Text(text = "Hello $name!")
+}
