+++ build.gradle.kts
@@ -68,7 +68,6 @@ buildscript {
 
 
 apply(from = "buildscripts/githooks.gradle")
-apply(from = "buildscripts/setup.gradle")
 apply(from = "buildscripts/versionsplugin.gradle")
 
 subprojects {
