+++ common/src/main/java/info/hridoydas/lifecanvas/common/components/TextFields.kt
@@ -1,0 +1,169 @@
+/*
+* MIT License
+*
+* Copyright (c) 2024 Hridoy Chandra Das
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in all
+* copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*
+*/
+package info.hridoydas.lifecanvas.common.components
+
+import androidx.annotation.StringRes
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Email
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusDirection
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import info.hridoydas.lifecanvas.common.R
+import info.hridoydas.lifecanvas.theme.LifeCanvasTheme
+
+@Composable
+fun AppTextField(
+    value: String,
+    @StringRes label: Int,
+    hint: String = "",
+    onValueChanged: (value: String) -> Unit,
+    isPasswordField: Boolean = false,
+    isClickOnly: Boolean = false,
+    onClick: () -> Unit = {},
+    leadingIcon: ImageVector? = null,
+    @StringRes error: Int? = null,
+    keyboardType: KeyboardType = KeyboardType.Text,
+    imeAction: ImeAction = ImeAction.Done,
+    onDone: () -> Unit = {},
+) {
+    val focusManager = LocalFocusManager.current
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(90.dp),
+    ) {
+        TextField(
+            modifier = Modifier
+                .fillMaxWidth()
+                .clickable {
+                    if (isClickOnly) {
+                        onClick()
+                    }
+                },
+            value = value,
+            onValueChange = { onValueChanged(it) },
+            singleLine = true,
+            isError = error != null,
+            readOnly = isClickOnly,
+            enabled = !isClickOnly,
+            supportingText = {
+                if (error != null) {
+                    Text(
+                        modifier = Modifier.fillMaxWidth(),
+                        text = stringResource(error),
+                        color = MaterialTheme.colorScheme.error,
+                    )
+                }
+            },
+            visualTransformation = if (isPasswordField) PasswordVisualTransformation() else VisualTransformation.None,
+            label = { Text(text = stringResource(label)) },
+            placeholder = { Text(text = hint) },
+            leadingIcon = {
+                leadingIcon?.let {
+                    Icon(it, hint, tint = MaterialTheme.colorScheme.onSurfaceVariant)
+                }
+            },
+            trailingIcon = {
+                if (error != null) {
+                    Icon(Icons.Filled.Info, "error", tint = MaterialTheme.colorScheme.error)
+                }
+            },
+            keyboardActions = KeyboardActions(
+                onDone = {
+                    focusManager.clearFocus()
+                    onDone()
+                },
+                onNext = { focusManager.moveFocus(FocusDirection.Down) },
+            ),
+            keyboardOptions = KeyboardOptions(
+                keyboardType = keyboardType,
+                imeAction = imeAction,
+            ),
+        )
+    }
+}
+
+@TemplatePreview
+@Composable
+fun AppTextFieldPreview() {
+    LifeCanvasTheme {
+        Surface {
+            TextField(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable {},
+                value = "",
+                onValueChange = { },
+                singleLine = true,
+                isError = true,
+                readOnly = false,
+                enabled = true,
+                supportingText = {
+                    Text(text = "Error Message or Supporting Message")
+                },
+                placeholder = { Text(text = "Hint") },
+            )
+        }
+    }
+}
+
+@TemplatePreview
+@Composable
+fun LoginScreenPreview() {
+    LifeCanvasTheme {
+        Surface(modifier = Modifier.fillMaxSize()) {
+            AppTextField(
+                value = "template@gmail.com",
+                label = R.string.email,
+                hint = "yourname@domain.com",
+                error = com.google.android.material.R.string.error_a11y_label,
+                leadingIcon = Icons.Filled.Email,
+                onValueChanged = { },
+                imeAction = ImeAction.Next,
+            )
+        }
+    }
+}
