+++ theme/src/main/java/info/hridoydas/lifecanvas/theme/Color.kt
@@ -1,0 +1,154 @@
+/*
+* MIT License
+*
+* Copyright (c) 2024 Hridoy Chandra Das
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in all
+* copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*
+*/
+package info.hridoydas.lifecanvas.theme
+import androidx.compose.ui.graphics.Color
+
+const val LIGHT_PRIMARY = 0xFF3F51B5
+const val LIGHT_ON_PRIMARY = 0xFFFFFFFF
+const val LIGHT_PRIMARY_CONTAINER = 0xFFFFFFFF
+const val LIGHT_ON_PRIMARY_CONTAINER = 0xFF000000
+const val LIGHT_SECONDARY = 0xFF303F9F
+const val LIGHT_ON_SECONDARY = 0xFFFFFFFF
+const val LIGHT_SECONDARY_CONTAINER = 0xFFFFFFFF
+const val LIGHT_ON_SECONDARY_CONTAINER = 0xFF000000
+const val LIGHT_TERTIARY = 0xFFC5CAE9
+const val LIGHT_ON_TERTIARY = 0xFF000000
+const val LIGHT_TERTIARY_CONTAINER = 0xFFFFFFFF
+const val LIGHT_ON_TERTIARY_CONTAINER = 0xFF000000
+const val LIGHT_ERROR = 0xFFB00020
+const val LIGHT_ERROR_CONTAINER = 0xFFFFFFFF
+const val LIGHT_ON_ERROR = 0xFFFFFFFF
+const val LIGHT_ON_ERROR_CONTAINER = 0xFF000000
+const val LIGHT_BACKGROUND = 0xFFFFFFFF
+const val LIGHT_ON_BACKGROUND = 0xFF000000
+const val LIGHT_SURFACE = 0xFFFFFFFF
+const val LIGHT_ON_SURFACE = 0xFF000000
+const val LIGHT_SURFACE_VARIANT = 0xFFE8EAF6
+const val LIGHT_ON_SURFACE_VARIANT = 0xFF000000
+const val LIGHT_OUTLINE = 0xFF000000
+const val LIGHT_INVERSE_ON_SURFACE = 0xFFFFFFFF
+const val LIGHT_INVERSE_SURFACE = 0xFF000000
+const val LIGHT_INVERSE_PRIMARY = 0xFF000000
+const val LIGHT_SHADOW = 0x14000000
+const val LIGHT_SURFACE_TINT = 0x0A000000
+const val LIGHT_OUTLINE_VARIANT = 0x1F000000
+const val LIGHT_SCRIM = 0x33000000
+
+const val DARK_PRIMARY = 0xFF7986CB
+const val DARK_ON_PRIMARY = 0xFFFFFFFF
+const val DARK_PRIMARY_CONTAINER = 0xFF303F9F
+const val DARK_ON_PRIMARY_CONTAINER = 0xFFFFFFFF
+const val DARK_SECONDARY = 0xFF303F9F
+const val DARK_ON_SECONDARY = 0xFFFFFFFF
+const val DARK_SECONDARY_CONTAINER = 0xFF3F51B5
+const val DARK_ON_SECONDARY_CONTAINER = 0xFFFFFFFF
+const val DARK_TERTIARY = 0xFFC5CAE9
+const val DARK_ON_TERTIARY = 0xFF000000
+const val DARK_TERTIARY_CONTAINER = 0xFFFFFFFF
+const val DARK_ON_TERTIARY_CONTAINER = 0xFF000000
+const val DARK_ERROR = 0xFFB00020
+const val DARK_ERROR_CONTAINER = 0xFFFFFFFF
+const val DARK_ON_ERROR = 0xFFFFFFFF
+const val DARK_ON_ERROR_CONTAINER = 0xFF000000
+const val DARK_BACKGROUND = 0xFF121212
+const val DARK_ON_BACKGROUND = 0xFFFFFFFF
+const val DARK_SURFACE = 0xFF121212
+const val DARK_ON_SURFACE = 0xFFFFFFFF
+const val DARK_SURFACE_VARIANT = 0xFF1E1E1E
+const val DARK_ON_SURFACE_VARIANT = 0xFFFFFFFF
+const val DARK_OUTLINE = 0xFF000000
+const val DARK_INVERSE_ON_SURFACE = 0xFFFFFFFF
+const val DARK_INVERSE_SURFACE = 0xFF000000
+const val DARK_INVERSE_PRIMARY = 0xFF000000
+const val DARK_SHADOW = 0x14000000
+const val DARK_SURFACE_TINT = 0x0A000000
+const val DARK_OUTLINE_VARIANT = 0x1F000000
+const val DARK_SCRIM = 0x33000000
+
+const val SEED = 0xFF3F51B5
+
+val template_theme_light_primary = Color(LIGHT_PRIMARY)
+val template_theme_light_onPrimary = Color(LIGHT_ON_PRIMARY)
+val template_theme_light_primaryContainer = Color(LIGHT_PRIMARY_CONTAINER)
+val template_theme_light_onPrimaryContainer = Color(LIGHT_ON_PRIMARY_CONTAINER)
+val template_theme_light_secondary = Color(LIGHT_SECONDARY)
+val template_theme_light_onSecondary = Color(LIGHT_ON_SECONDARY)
+val template_theme_light_secondaryContainer = Color(LIGHT_SECONDARY_CONTAINER)
+val template_theme_light_onSecondaryContainer = Color(LIGHT_ON_SECONDARY_CONTAINER)
+val template_theme_light_tertiary = Color(LIGHT_TERTIARY)
+val template_theme_light_onTertiary = Color(LIGHT_ON_TERTIARY)
+val template_theme_light_tertiaryContainer = Color(LIGHT_TERTIARY_CONTAINER)
+val template_theme_light_onTertiaryContainer = Color(LIGHT_ON_TERTIARY_CONTAINER)
+val template_theme_light_error = Color(LIGHT_ERROR)
+val template_theme_light_errorContainer = Color(LIGHT_ERROR_CONTAINER)
+val template_theme_light_onError = Color(LIGHT_ON_ERROR)
+val template_theme_light_onErrorContainer = Color(LIGHT_ON_ERROR_CONTAINER)
+val template_theme_light_background = Color(LIGHT_BACKGROUND)
+val template_theme_light_onBackground = Color(LIGHT_ON_BACKGROUND)
+val template_theme_light_surface = Color(LIGHT_SURFACE)
+val template_theme_light_onSurface = Color(LIGHT_ON_SURFACE)
+val template_theme_light_surfaceVariant = Color(LIGHT_SURFACE_VARIANT)
+val template_theme_light_onSurfaceVariant = Color(LIGHT_ON_SURFACE_VARIANT)
+val template_theme_light_outline = Color(LIGHT_OUTLINE)
+val template_theme_light_inverseOnSurface = Color(LIGHT_INVERSE_ON_SURFACE)
+val template_theme_light_inverseSurface = Color(LIGHT_INVERSE_SURFACE)
+val template_theme_light_inversePrimary = Color(LIGHT_INVERSE_PRIMARY)
+val template_theme_light_shadow = Color(LIGHT_SHADOW)
+val template_theme_light_surfaceTint = Color(LIGHT_SURFACE_TINT)
+val template_theme_light_outlineVariant = Color(LIGHT_OUTLINE_VARIANT)
+val template_theme_light_scrim = Color(LIGHT_SCRIM)
+
+val template_theme_dark_primary = Color(DARK_PRIMARY)
+val template_theme_dark_onPrimary = Color(DARK_ON_PRIMARY)
+val template_theme_dark_primaryContainer = Color(DARK_PRIMARY_CONTAINER)
+val template_theme_dark_onPrimaryContainer = Color(DARK_ON_PRIMARY_CONTAINER)
+val template_theme_dark_secondary = Color(DARK_SECONDARY)
+val template_theme_dark_onSecondary = Color(DARK_ON_SECONDARY)
+val template_theme_dark_secondaryContainer = Color(DARK_SECONDARY_CONTAINER)
+val template_theme_dark_onSecondaryContainer = Color(DARK_ON_SECONDARY_CONTAINER)
+val template_theme_dark_tertiary = Color(DARK_TERTIARY)
+val template_theme_dark_onTertiary = Color(DARK_ON_TERTIARY)
+val template_theme_dark_tertiaryContainer = Color(DARK_TERTIARY_CONTAINER)
+val template_theme_dark_onTertiaryContainer = Color(DARK_ON_TERTIARY_CONTAINER)
+val template_theme_dark_error = Color(DARK_ERROR)
+val template_theme_dark_errorContainer = Color(DARK_ERROR_CONTAINER)
+val template_theme_dark_onError = Color(DARK_ON_ERROR)
+val template_theme_dark_onErrorContainer = Color(DARK_ON_ERROR_CONTAINER)
+val template_theme_dark_background = Color(DARK_BACKGROUND)
+val template_theme_dark_onBackground = Color(DARK_ON_BACKGROUND)
+val template_theme_dark_surface = Color(DARK_SURFACE)
+val template_theme_dark_onSurface = Color(DARK_ON_SURFACE)
+val template_theme_dark_surfaceVariant = Color(DARK_SURFACE_VARIANT)
+val template_theme_dark_onSurfaceVariant = Color(DARK_ON_SURFACE_VARIANT)
+val template_theme_dark_outline = Color(DARK_OUTLINE)
+val template_theme_dark_inverseOnSurface = Color(DARK_INVERSE_ON_SURFACE)
+val template_theme_dark_inverseSurface = Color(DARK_INVERSE_SURFACE)
+val template_theme_dark_inversePrimary = Color(DARK_INVERSE_PRIMARY)
+val template_theme_dark_shadow = Color(DARK_SHADOW)
+val template_theme_dark_surfaceTint = Color(DARK_SURFACE_TINT)
+val template_theme_dark_outlineVariant = Color(DARK_OUTLINE_VARIANT)
+val template_theme_dark_scrim = Color(DARK_SCRIM)
+
+val seed = Color(SEED)
